Simplify code

+++ custom_news_feed/custom_news_feed.py
@@ -551,9 +551,12 @@ class CustomNewsFeed:
     def remove_deprecated_hashfiles(self):
         """Removes hash files that are not in the current news anymore"""
         for file_path in os.listdir(self.settingspath):
-            if file_path != self.previous_news_name and os.path.isfile(os.path.join(self.settingspath, file_path)):
-                if all([article["Hash"] != file_path for article in self.news["NewsArticles"]]):
-                    self.delete_hashfile(file_path)
+            if (
+                file_path != self.previous_news_name and 
+                os.path.isfile(os.path.join(self.settingspath, file_path)) and
+                all([article["Hash"] != file_path for article in self.news["NewsArticles"]])
+            ):
+                self.delete_hashfile(file_path)
 
     def create_tab_widget(self, tab):
         """ Creates a layout for the news articles and adds it to the tab """
